+++ Core/Library/Extensions/TypeIsIntegralExtensions.cs
@@ -2,24 +2,20 @@
 
 namespace Xamarin.Realm.Service.Extensions
 {
-    public static class TypeIsIntegralExtensions
+    internal static class TypeIsIntegralExtensions
     {
-        public static bool IsIntegral(this Type type)
+        internal static bool IsIntegral(this Type type)
         {
             if (type == null)
             {
                 return false;
             }
 
-            if (type == typeof(sbyte) ||
-                type == typeof(byte) ||
+            if (type == typeof(byte) ||
                 type == typeof(char) ||
                 type == typeof(short) ||
-                type == typeof(ushort) ||
                 type == typeof(int) ||
-                type == typeof(uint) ||
-                type == typeof(long) ||
-                type == typeof(ulong))
+                type == typeof(long))
             {
                 return true;
             }